bullets v2

+++ core/src/main/java/com/untilldown/Model/Enums/WeaponType.java
@@ -13,14 +13,14 @@ public enum WeaponType {
     ;
 
     private final String name;
-    private final double damage;
+    private final float damage;
     private final int projectile;
     private final double reloadTime;
     private final int maxAmmo;
     private final String imagePath;
     private final int numOfBullets;
 
-    WeaponType(String name, double damage, int projectile, double reloadTime, int maxAmmo, String path, int numOfBullets) {
+    WeaponType(String name, float damage, int projectile, double reloadTime, int maxAmmo, String path, int numOfBullets) {
         this.name = name;
         this.damage = damage;
         this.projectile = projectile;
@@ -42,7 +42,7 @@ public enum WeaponType {
         return name;
     }
 
-    public double getDamage() {
+    public float getDamage() {
         return damage;
     }
